==> exp2/scripts/03_analyze_eyedata.R <==
#################
## LOAD/preprocess the data
#################
suppressWarnings(
  suppressPackageStartupMessages({
    library("tidyverse")
    source("../global_fns.R")
  }))

message("    Preparing eye data...")

.old_obj <- load("data_images/01_preprocess.rda")

tresp <- main_data %>%
  filter(!is.na(Misspec)) %>%
  select(SessionID, Series, Misspec, Onset) %>%
  inner_join(tmatchup %>%
             select(SessionID, RespID, Series, shift_dir, congruency),
             c("SessionID", "Series")) %>%
  select(SessionID,
         RespID,
         congruency, 
         shift_dir,
         misspec = Misspec,
         onset = Onset)

## point of gaze (POG) data
pog2 <- pog %>%
  filter(FrameID <= 500L) %>%
  select(RespID, FrameID, Name, Region)

#################
## count the number of observations for each combination of
## SessionID, congruency, shift_dir, FrameID
#################

binned <- tresp %>%
  inner_join(pog2, "RespID") %>%
  group_by(SessionID, congruency, shift_dir, FrameID, Region, .drop = FALSE) %>%
  summarise(Y = n()) %>%
  group_by(SessionID, congruency, shift_dir, FrameID) %>%
  mutate(N = sum(Y), p = Y / N) %>%
  ungroup()

bins_means <- agg_up2(binned, Y, N, congruency, shift_dir, FrameID, Region)

nmc <- 1000L
message("    Running ", nmc, " bootstrap samples (takes a long time)...")

mx <- replicate(nmc, boot_once(binned, SessionID) %>%
  agg_up2(Y, N, congruency, shift_dir, FrameID, Region) %>%
  pull(p))

conf <- apply(mx, 1, quantile, probs = c(.025, .975))

bins_means[["lower"]] = conf["2.5%", ]
bins_means[["upper"]] = conf["97.5%", ]

message("    Creating probability plot...")

exp2_probplot <- bins_means %>%
  filter(Region != "blank") %>%
  mutate(ms = round(1000 * ((FrameID - 1) / 250))) %>%
  ggplot(aes(ms, p, linetype = congruency)) +
  geom_ribbon(aes(fill = Region, ymin = lower, ymax = upper),
	      alpha = .2) +
  geom_smooth(aes(color = Region), se = FALSE, span = .05, alpha = .6) +
  facet_wrap(~shift_dir, nrow = 2) +
  scale_fill_manual(values = cbPalette) +
  scale_color_manual(values = cbPalette) +
  labs(x = "Time from display onset (ms)",
       y = "Gaze probability",
       linetype = "Congruency")



message("    Writing data_images/03_analyze_eyedata.rda...")
save(list = c("bins_means", "binned", "nmc", "exp2_probplot"),
     file = "data_images/03_analyze_eyedata.rda")
